refactor(frontend): move all loaders components in specific folder
