+++ src/frontend/src/lib/components/loaders/Loader.svelte
