+++ src/frontend/src/lib/components/loaders/LoaderMetamask.svelte
