+++ src/frontend/src/lib/components/loaders/LoaderWallets.svelte
