+++ src/frontend/src/lib/components/loaders/Loaders.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
 	import LoaderBalances from '$icp-eth/components/core/LoaderBalances.svelte';
-	import Loader from '$lib/components/core/Loader.svelte';
-	import LoaderMetamask from '$lib/components/core/LoaderMetamask.svelte';
-	import LoaderWallets from '$lib/components/core/LoaderWallets.svelte';
+	import Loader from '$lib/components/loaders/Loader.svelte';
+	import LoaderMetamask from '$lib/components/loaders/LoaderMetamask.svelte';
+	import LoaderWallets from '$lib/components/loaders/LoaderWallets.svelte';
 	import ExchangeWorker from '$lib/components/exchange/ExchangeWorker.svelte';
 	import AddressGuard from '$lib/components/guard/AddressGuard.svelte';
 </script>
