+++ src/frontend/src/routes/(app)/+layout.svelte
@@ -5,7 +5,7 @@
 	import { page } from '$app/stores';
 	import AuthGuard from '$lib/components/auth/AuthGuard.svelte';
 	import Footer from '$lib/components/core/Footer.svelte';
-	import LoadersGuard from '$lib/components/core/LoadersGuard.svelte';
+	import LoadersGuard from '$lib/components/loaders/LoadersGuard.svelte';
 	import Modals from '$lib/components/core/Modals.svelte';
 	import DappsCarousel from '$lib/components/dapps/DappsCarousel.svelte';
 	import Header from '$lib/components/hero/Header.svelte';
